+++ minijinja/src/filters.rs
@@ -5,8 +5,8 @@
 //! invokes the filter `filter` with the arguments `42` and `23`.
 //!
 //! MiniJinja comes with some built-in filters that are listed below. To create a
-//! custom filter write a function that takes at least an
-//! [`&Environment`](crate::Environment) and value argument, then register it
+//! custom filter write a function that takes at least a
+//! [`&State`](crate::State) and value argument, then register it
 //! with [`add_filter`](crate::Environment::add_filter).
 //!
 //! ## Custom Filters
@@ -16,9 +16,9 @@
 //! value and replaces whitespace with dashes and converts it to lowercase:
 //!
 //! ```
-//! # use minijinja::{Environment, Error};
+//! # use minijinja::{Environment, State, Error};
 //! # let mut env = Environment::new();
-//! fn slugify(env: &Environment, value: String) -> Result<String, Error> {
+//! fn slugify(_state: &State, value: String) -> Result<String, Error> {
 //!     Ok(value.to_lowercase().split_whitespace().collect::<Vec<_>>().join("-"))
 //! }
 //!
@@ -29,13 +29,12 @@
 //! [`FunctionArgs`](crate::value::FunctionArgs) and [`Into`] traits.
 use std::collections::BTreeMap;
 
-use crate::environment::Environment;
 use crate::error::Error;
 use crate::utils::HtmlEscape;
 use crate::value::{ArgType, FunctionArgs, RcType, Value};
+use crate::vm::State;
 
-type FilterFunc =
-    dyn Fn(&Environment, Value, Vec<Value>) -> Result<Value, Error> + Sync + Send + 'static;
+type FilterFunc = dyn Fn(&State, Value, Vec<Value>) -> Result<Value, Error> + Sync + Send + 'static;
 
 #[derive(Clone)]
 pub(crate) struct BoxedFilter(RcType<FilterFunc>);
@@ -43,19 +42,19 @@ pub(crate) struct BoxedFilter(RcType<FilterFunc>);
 /// A utility trait that represents filters.
 pub trait Filter<V = Value, Rv = Value, Args = Vec<Value>>: Send + Sync + 'static {
     /// Applies a filter to value with the given arguments.
-    fn apply_to(&self, env: &Environment, value: V, args: Args) -> Result<Rv, Error>;
+    fn apply_to(&self, state: &State, value: V, args: Args) -> Result<Rv, Error>;
 }
 
 macro_rules! tuple_impls {
     ( $( $name:ident )* ) => {
         impl<Func, V, Rv, $($name),*> Filter<V, Rv, ($($name,)*)> for Func
         where
-            Func: Fn(&Environment, V, $($name),*) -> Result<Rv, Error> + Send + Sync + 'static
+            Func: Fn(&State, V, $($name),*) -> Result<Rv, Error> + Send + Sync + 'static
         {
-            fn apply_to(&self, env: &Environment, value: V, args: ($($name,)*)) -> Result<Rv, Error> {
+            fn apply_to(&self, state: &State, value: V, args: ($($name,)*)) -> Result<Rv, Error> {
                 #[allow(non_snake_case)]
                 let ($($name,)*) = args;
-                (self)(env, value, $($name,)*)
+                (self)(state, value, $($name,)*)
             }
         }
     };
@@ -77,9 +76,9 @@ impl BoxedFilter {
         Args: FunctionArgs,
     {
         BoxedFilter(RcType::new(
-            move |env, value, args| -> Result<Value, Error> {
+            move |state, value, args| -> Result<Value, Error> {
                 f.apply_to(
-                    env,
+                    state,
                     ArgType::from_value(Some(value))?,
                     FunctionArgs::from_values(args)?,
                 )
@@ -89,13 +88,8 @@ impl BoxedFilter {
     }
 
     /// Applies the filter to a value and argument.
-    pub fn apply_to(
-        &self,
-        env: &Environment,
-        value: Value,
-        args: Vec<Value>,
-    ) -> Result<Value, Error> {
-        (self.0)(env, value, args)
+    pub fn apply_to(&self, state: &State, value: Value, args: Vec<Value>) -> Result<Value, Error> {
+        (self.0)(state, value, args)
     }
 }
 
@@ -122,7 +116,7 @@ pub(crate) fn get_builtin_filters() -> BTreeMap<&'static str, BoxedFilter> {
 }
 
 /// Marks a value as safe.  This converts it into a string.
-pub fn safe(_env: &Environment, v: String) -> Result<Value, Error> {
+pub fn safe(_state: &State, v: String) -> Result<Value, Error> {
     // TODO: this ideally understands which type of escaping is in use
     Ok(Value::from_safe_string(v))
 }
@@ -130,7 +124,7 @@ pub fn safe(_env: &Environment, v: String) -> Result<Value, Error> {
 /// HTML escapes a string.
 ///
 /// By default this filter is also registered under the alias `e`.
-pub fn escape(_env: &Environment, v: Value) -> Result<Value, Error> {
+pub fn escape(_state: &State, v: Value) -> Result<Value, Error> {
     // TODO: this ideally understands which type of escaping is in use
     if v.is_safe() {
         Ok(v)
@@ -153,24 +147,19 @@ mod builtins {
 
     /// Converts a value to uppercase.
     #[cfg_attr(docsrs, doc(cfg(feature = "builtin_filters")))]
-    pub fn upper(_env: &Environment, v: String) -> Result<String, Error> {
+    pub fn upper(_state: &State, v: String) -> Result<String, Error> {
         Ok(v.to_uppercase())
     }
 
     /// Converts a value to lowercase.
     #[cfg_attr(docsrs, doc(cfg(feature = "builtin_filters")))]
-    pub fn lower(_env: &Environment, v: String) -> Result<String, Error> {
+    pub fn lower(_state: &State, v: String) -> Result<String, Error> {
         Ok(v.to_lowercase())
     }
 
     /// Does a string replace.
     #[cfg_attr(docsrs, doc(cfg(feature = "builtin_filters")))]
-    pub fn replace(
-        _env: &Environment,
-        v: String,
-        from: String,
-        to: String,
-    ) -> Result<String, Error> {
+    pub fn replace(_state: &State, v: String, from: String, to: String) -> Result<String, Error> {
         Ok(v.replace(&from, &to))
     }
 
@@ -178,7 +167,7 @@ mod builtins {
     ///
     /// By default this filter is also registered under the alias `count`.
     #[cfg_attr(docsrs, doc(cfg(feature = "builtin_filters")))]
-    pub fn length(_env: &Environment, v: Value) -> Result<Value, Error> {
+    pub fn length(_state: &State, v: Value) -> Result<Value, Error> {
         v.len().map(Value::from).ok_or_else(|| {
             Error::new(
                 ErrorKind::ImpossibleOperation,
@@ -189,7 +178,7 @@ mod builtins {
 
     /// Dict sorting functionality.
     #[cfg_attr(docsrs, doc(cfg(feature = "builtin_filters")))]
-    pub fn dictsort(_env: &Environment, v: Value) -> Result<Value, Error> {
+    pub fn dictsort(_state: &State, v: Value) -> Result<Value, Error> {
         let mut pairs = v.try_into_pairs()?;
         pairs.sort_by(|a, b| a.partial_cmp(b).unwrap_or(Ordering::Equal));
         Ok(Value::from(pairs))
@@ -197,7 +186,7 @@ mod builtins {
 
     /// Reverses a list or string
     #[cfg_attr(docsrs, doc(cfg(feature = "builtin_filters")))]
-    pub fn reverse(_env: &Environment, v: Value) -> Result<Value, Error> {
+    pub fn reverse(_state: &State, v: Value) -> Result<Value, Error> {
         if let Some(Primitive::Str(s)) = v.as_primitive() {
             Ok(Value::from(s.chars().rev().collect::<String>()))
         } else if matches!(v.kind(), ValueKind::Seq) {
@@ -214,7 +203,7 @@ mod builtins {
 
     /// Trims a value
     #[cfg_attr(docsrs, doc(cfg(feature = "builtin_filters")))]
-    pub fn trim(_env: &Environment, s: String, chars: Option<String>) -> Result<String, Error> {
+    pub fn trim(_state: &State, s: String, chars: Option<String>) -> Result<String, Error> {
         match chars {
             Some(chars) => {
                 let chars = chars.chars().collect::<Vec<_>>();
@@ -226,7 +215,7 @@ mod builtins {
 
     /// Joins a sequence by a character
     #[cfg_attr(docsrs, doc(cfg(feature = "builtin_filters")))]
-    pub fn join(_env: &Environment, val: Value, joiner: Option<String>) -> Result<String, Error> {
+    pub fn join(_state: &State, val: Value, joiner: Option<String>) -> Result<String, Error> {
         if val.is_undefined() || val.is_none() {
             return Ok(String::new());
         }
@@ -267,7 +256,7 @@ mod builtins {
     ///
     /// By default this filter is also registered under the alias `d`.
     #[cfg_attr(docsrs, doc(cfg(feature = "builtin_filters")))]
-    pub fn default(_: &Environment, value: Value, other: Option<Value>) -> Result<Value, Error> {
+    pub fn default(_: &State, value: Value, other: Option<Value>) -> Result<Value, Error> {
         Ok(if value.is_undefined() {
             other.unwrap_or_else(|| Value::from(""))
         } else {
@@ -277,14 +266,16 @@ mod builtins {
 
     #[test]
     fn test_basics() {
-        fn test(_: &Environment, a: u32, b: u32) -> Result<u32, Error> {
+        fn test(_: &State, a: u32, b: u32) -> Result<u32, Error> {
             Ok(a + b)
         }
 
-        let env = Environment::new();
+        let env = crate::Environment::new();
+        let ctx = crate::vm::Context::default();
+        let state = State::from_env_and_context(&env, &ctx);
         let bx = BoxedFilter::new(test);
         assert_eq!(
-            bx.apply_to(&env, Value::from(23), vec![Value::from(42)])
+            bx.apply_to(&state, Value::from(23), vec![Value::from(42)])
                 .unwrap(),
             Value::from(65)
         );
@@ -292,7 +283,7 @@ mod builtins {
 
     #[test]
     fn test_optional_args() {
-        fn add(_: &Environment, val: u32, a: u32, b: Option<u32>) -> Result<u32, Error> {
+        fn add(_: &State, val: u32, a: u32, b: Option<u32>) -> Result<u32, Error> {
             let mut sum = val + a;
             if let Some(b) = b {
                 sum += b;
@@ -300,16 +291,18 @@ mod builtins {
             Ok(sum)
         }
 
-        let env = Environment::new();
+        let env = crate::Environment::new();
+        let ctx = crate::vm::Context::default();
+        let state = State::from_env_and_context(&env, &ctx);
         let bx = BoxedFilter::new(add);
         assert_eq!(
-            bx.apply_to(&env, Value::from(23), vec![Value::from(42)])
+            bx.apply_to(&state, Value::from(23), vec![Value::from(42)])
                 .unwrap(),
             Value::from(65)
         );
         assert_eq!(
             bx.apply_to(
-                &env,
+                &state,
                 Value::from(23),
                 vec![Value::from(42), Value::UNDEFINED]
             )
@@ -317,8 +310,12 @@ mod builtins {
             Value::from(65)
         );
         assert_eq!(
-            bx.apply_to(&env, Value::from(23), vec![Value::from(42), Value::from(1)])
-                .unwrap(),
+            bx.apply_to(
+                &state,
+                Value::from(23),
+                vec![Value::from(42), Value::from(1)]
+            )
+            .unwrap(),
             Value::from(66)
         );
     }